+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Certbot adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 
-*
+* Turn off session tickets for apache plugin by default
 
 ### Changed
 
@@ -23,7 +23,6 @@ More details about these changes can be found on our GitHub repo.
 ### Added
 
 * Turn off session tickets for nginx plugin by default
-* Turn off session tickets for apache plugin by default
 * Added missing error types from RFC8555 to acme
 
 ### Changed